Refactor data access and view models

+++ composeApp/src/commonMain/kotlin/org/example/project/model/dto/ItemDto.kt
@@ -14,7 +14,7 @@ data class ItemDto(
     @Serializable(with = OffsetDateTimeSerializer::class) val startDate: OffsetDateTime? = null,
     @Serializable(with = OffsetDateTimeSerializer::class) val finishDate: OffsetDateTime? = null,
     @Serializable(with = OffsetDateTimeSerializer::class) val shipDate: OffsetDateTime? = null,
-    val timesUploaded: Int? = null,
+    val timesUploaded: Long? = null,
     @Contextual val images: List<ImageDto>? = emptyList(),
     val hashtags: List<String>? = emptyList(),
     @Contextual val person: PersonDto? = null,
